Add adapter for CrateDB
It is not suitable for merging into mainline, because it modifies the
data model inappropriately for other databases. The missing details
would need to be put into the dialect somehow, to make it compatible
with a vanilla SQLAlchemy application. C'est la vie.

The patch includes a few monkeypatches and polyfills, which should be
upstreamed to crate-python and/or cratedb-toolkit.